Updating the client interface 27.06.2025

+++ server/src/main/java/ru/avdonin/server/service/list/AvatarFtpService.java
@@ -18,7 +18,7 @@ public class AvatarFtpService extends AbstractFtpService {
                 "default",
                 "default-avatar.png",
                 "/user_avatars",
-                (Integer) Constants.COMPRESSION_AVATAR.getValue(),
-                (Integer) Constants.COMPRESSION_AVATAR.getValue());
+                Constants.COMPRESSION_AVATAR.getValue(),
+                Constants.COMPRESSION_AVATAR.getValue());
     }
 }
